% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_multitrack_lines.R
\name{add_multitrack_lines}
\alias{add_multitrack_lines}
\title{Add vertical lines}
\usage{
add_multitrack_lines(
  TRKS,
  dat,
  snp_groups = c("Lead", "UCS", "Consensus"),
  line_alpha = 1,
  line_size = 0.3,
  remove_duplicated_UCS_Consensus = TRUE,
  verbose = FALSE
)
}
\description{
Adds vertical lines indicates key SNPs
}
\seealso{
Other plot: 
\code{\link{dot_summary}()},
\code{\link{get_window_limits}()},
\code{\link{get_window_suffix}()}
}
\concept{plot}
\keyword{internal}
