% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_snp_labels.R
\name{add_snp_labels}
\alias{add_snp_labels}
\title{Add SNP labels}
\usage{
add_snp_labels(
  snp_plot,
  dat,
  labels_subset = c("CS"),
  yvar = "PP",
  genomic_units = "Mb",
  grouping_vars = c("SNP", "Method"),
  remove_duplicates = FALSE,
  melt_methods = TRUE,
  show.legend = TRUE
)
}
\arguments{
\item{dat}{GWAS summary statistics subset to query the LD panel with.}
}
\description{
Add SNP labels for Manhattan plot.
Support function for \link[echoplot]{plot_locus}.
}
\keyword{internal}
