% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_summary.R
\name{dot_summary}
\alias{dot_summary}
\title{Multi-fine-map summary dot plot}
\usage{
dot_summary(dat, PP_threshold = 0.95, show_plot = T)
}
\description{
Multi-fine-map summary dot plot
}
\examples{
data("BST1")
gp <- dot_summary(dat=BST1)
}
\seealso{
Other plot: 
\code{\link{add_multitrack_lines}()},
\code{\link{get_window_limits}()},
\code{\link{get_window_suffix}()}
}
\concept{plot}
