% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_summary.R
\name{dot_summary_plot}
\alias{dot_summary_plot}
\title{Dot summary plot}
\usage{
dot_summary_plot(dat, PP_threshold = 0.95, show_plot = TRUE)
}
\description{
Multi-fine-map summary dot plot of all lead, UCS and/or Consensus SNPs.
}
\examples{
dat <- echodata::BST1
gp <- echoplot::dot_summary_plot(dat=dat)
}
\seealso{
Other plot: 
\code{\link{add_multitrack_lines}()},
\code{\link{get_window_limits}()},
\code{\link{get_window_suffix}()}
}
\concept{plot}
