% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_histogram_height.R
\name{get_max_histogram_height}
\alias{get_max_histogram_height}
\title{Get max histogram height}
\usage{
get_max_histogram_height(gg, round_to = NULL, verbose = TRUE)
}
\description{
Get the max height in a histogram plot.
}
\keyword{internal}
