% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_palettes.R
\name{get_palettes}
\alias{get_palettes}
\title{Get palettes}
\usage{
get_palettes(preferred = c("Spectral", "BrBG", "PiYG", "PuOr"))
}
\arguments{
\item{preferred}{Preferred palette names.}
}
\value{
Vector of palette names.
}
\description{
Get a unique set of palettes, ordered by preference.
}
\keyword{internal}
