% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transcripts.R
\name{get_transcripts}
\alias{get_transcripts}
\title{Get transcripts}
\source{
\code{
#### Alternative approaches I've tried ####

gr.snp <- GenomicRanges::makeGRangesFromDataFrame(echolocatoR::LRRK2, keep.extra.columns = T, seqnames.field = "CHR", start.field = "POS", end.field = "POS")

# Warning:: MUST load the full package bc
# it loads other necessary packages into the namespace.

BiocManager::install("Homo.sapiens")
library(Homo.sapiens)
txdb <- Homo.sapiens::Homo.sapiens
GenomicFeatures::genes(txdb)

BiocManager::install("TxDb.Hsapiens.UCSC.hg19.knownGene")
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene
GenomicFeatures::genes(txdb)
}
}
\usage{
get_transcripts(
  gr.snp,
  max_transcripts = 1,
  remove_pseudogenes = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{Print messages.}
}
\description{
Get transcript positions for all genes in a range.
}
\keyword{internal}
