% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tx_biotypes.R
\name{get_tx_biotypes}
\alias{get_tx_biotypes}
\title{Get transcript biotypes}
\usage{
get_tx_biotypes(tx_biotypes = NULL, as_filter = TRUE, verbose = TRUE)
}
\arguments{
\item{tx_biotypes}{Transcript biotypes to include in the gene model track. 
By default (\code{NULL}), all transcript biotypes will be included.
See \link[echoplot]{get_transcripts_biotypes} for a full list of 
all available biotypes}

\item{as_filter}{Return results as an
\link[AnnotationFilter]{TxBiotypeFilter} (default: \code{TRUE}). 
Otherwise, a character vector will be returned.}

\item{verbose}{Print messages.}
}
\value{
Character vector of valid transcript biotypes.
}
\description{
List all available transcript biotypes in the 
\link[EnsDb.Hsapiens.v75]{EnsDb.Hsapiens.v75} database.
}
\examples{
tx_filter <- get_transcripts_biotypes()
}
