% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_window_limits.R
\name{get_window_limits}
\alias{get_window_limits}
\title{Get window size limits for plot}
\usage{
get_window_limits(
  dat,
  index_as_center = T,
  plot.zoom = NULL,
  genomic_units = "Mb",
  verbose = T
)
}
\description{
Get window size limits for plot
}
\examples{
data("BST1");
xlims <- get_window_limits(dat=BST1, plot.zoom=50000)
xlims <- get_window_limits(dat=BST1, plot.zoom="all")
xlims <- get_window_limits(dat=BST1, plot.zoom="5x")
}
\seealso{
Other plot: 
\code{\link{add_multitrack_lines}()},
\code{\link{dot_summary}()},
\code{\link{get_window_suffix}()}
}
\concept{plot}
\keyword{internal}
