% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_window_suffix.R
\name{get_window_suffix}
\alias{get_window_suffix}
\title{Determine the plot suffix indicating its window size}
\usage{
get_window_suffix(dat, plot.zoom, verbose = T)
}
\description{
Determine the plot suffix indicating its window size
}
\examples{
data("BST1")
window_suffix <- get_window_suffix(dat=BST1, plot.zoom=1000)
window_suffix <- get_window_suffix(dat=BST1, plot.zoom=NULL)
window_suffix <- get_window_suffix(dat=BST1, plot.zoom="all")
window_suffix <- get_window_suffix(dat=BST1, plot.zoom="2x")
}
\seealso{
Other plot: 
\code{\link{add_multitrack_lines}()},
\code{\link{dot_summary}()},
\code{\link{get_window_limits}()}
}
\concept{plot}
\keyword{internal}
