% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_namespace.R
\name{list_namespace}
\alias{list_namespace}
\title{List namespace}
\usage{
list_namespace(packages, status_filter = NULL, verbose = TRUE)
}
\arguments{
\item{packages}{A character vector of package names.}

\item{status_filter}{Filter by namespace status (e.g. "export","import").
If \code{NULL} (default), will return all namespace objects.}

\item{verbose}{Print messages.}
}
\value{
A \link[data.table]{data.table}.
}
\description{
List all objects (e.g. functions, data) 
in the namespace one or more R packages.
}
\examples{
nm_dt <- list_namespace(packages=c("data.table","pals"))
}
