% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_locus.R
\name{plot_locus}
\alias{plot_locus}
\title{Generate a locus plot}
\usage{
plot_locus(
  dat,
  locus_dir,
  LD_matrix = NULL,
  LD_reference = NULL,
  dataset_type = "GWAS",
  color_r2 = TRUE,
  method_list = c("ABF", "FINEMAP", "SUSIE", "POLYFUN_SUSIE"),
  track_order = NULL,
  track_heights = NULL,
  plot_full_window = TRUE,
  dot_summary = FALSE,
  QTL_prefixes = NULL,
  mean.PP = TRUE,
  PP_threshold = 0.95,
  consensus_threshold = 2,
  sig_cutoff = 5e-08,
  gene_track = TRUE,
  point_size = 1,
  point_alpha = 0.6,
  snp_group_lines = c("Lead", "UCS", "Consensus"),
  xtext = FALSE,
  show.legend_genes = TRUE,
  XGR_libnames = NULL,
  n_top_xgr = 5,
  Roadmap = FALSE,
  Roadmap_query = NULL,
  n_top_roadmap = 7,
  annot_overlap_threshold = 5,
  zoom_exceptions_str = "*full window$|zoom_polygon",
  Nott_epigenome = FALSE,
  Nott_regulatory_rects = TRUE,
  Nott_show_placseq = TRUE,
  Nott_binwidth = 200,
  Nott_bigwig_dir = NULL,
  save_plot = TRUE,
  show_plot = TRUE,
  genomic_units = "Mb",
  strip.text.y.angle = 0,
  max_transcripts = 1,
  plot.zoom = c("1x"),
  dpi = 300,
  height = 12,
  width = 10,
  plot_format = "jpg",
  save_RDS = FALSE,
  return_list = FALSE,
  conda_env = "echoR",
  nThread = 1,
  verbose = TRUE
)
}
\description{
Generate a locus-specific plot with multiple selectable tracks.
Users can also generate multiple zoomed in views of the plot at
 multiple resolutions.
}
\examples{
dat<- echodata::BST1
LD_matrix <- echodata::BST1_LD_matrix
locus_dir <- file.path(tempdir(),echodata::locus_dir)

plot_list <- echoplot::plot_locus(dat = dat,
                                  locus_dir=locus_dir,
                                  LD_matrix=LD_matrix)
}
