% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_locus_multi.R
\name{plot_locus_multi}
\alias{plot_locus_multi}
\title{Plot multi-GWAS/QTL or multi-ancestry fine-mapping results}
\usage{
plot_locus_multi(
  dat_ls,
  LD_ls,
  locus_dir,
  conditions = names(dat_ls),
  show_plot = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dat_ls}{A named list of \link[data.table]{data.table}s generated by
\code{echolocatoR::finemap_loci}.}

\item{LD_ls}{A named list of link disequilibrium (LD) matrices 
(one per item in \code{dat_ls}).}

\item{locus_dir}{Storage directory to use.}

\item{conditions}{Conditions to group \code{dat_ls} by.
The length of \code{conditions} must equal the number of
 items in \code{dat_ls}.}

\item{show_plot}{Print plot to screen.}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[=plot_locus]{plot_locus}}
  \describe{
    \item{\code{dataset_type}}{Dataset type (e.g. "GWAS" or "eQTL").}
    \item{\code{color_r2}}{Whether to color data points (SNPs)
by how strongly they correlate with the lead SNP 
(i.g. LD measured in terms of r2).}
    \item{\code{finemap_methods}}{Fine-mapping methods to plot tracks for, 
where the y-axis show the Posterior Probabilities (PP) 
of each SNP being causal.}
    \item{\code{track_order}}{The order in which tracks should appear
(from top to bottom).}
    \item{\code{track_heights}}{The height of each track (from top to bottom).}
    \item{\code{plot_full_window}}{Include a track with a Manhattan plot of the full
GWAS/eQTL locus (not just the zoomed-in portion).}
    \item{\code{dot_summary}}{Include a dot-summary plot that highlights the
Lead, Credible Set, and Consensus SNPs.}
    \item{\code{qtl_suffixes}}{If columns with QTL data is included in \code{dat},
you can indicate which columns those are with one or more string suffixes
(e.g. \code{qtl_suffixes=c(".eQTL1",".eQTL2")} to use the columns 
"P.QTL1", "Effect.QTL1", "P.QTL2", "Effect.QTL2").}
    \item{\code{mean.PP}}{Include a track showing mean Posterior Probabilities (PP) 
averaged across all fine-mapping methods.}
    \item{\code{sig_cutoff}}{Filters out SNPs to plot based on an (uncorrected)
p-value significance cutoff.}
    \item{\code{gene_track}}{Include a track showing gene bodies.}
    \item{\code{max_transcripts}}{Maximum number of transcripts per gene.}
    \item{\code{tx_biotypes}}{Transcript biotypes to include in the gene model track. 
By default (\code{NULL}), all transcript biotypes will be included.
See \link[echoplot]{get_tx_biotypes} for a full list of 
all available biotypes}
    \item{\code{point_size}}{Size of each data point.}
    \item{\code{point_alpha}}{Opacity of each data point.}
    \item{\code{snp_group_lines}}{Include vertical lines to help highlight 
SNPs belonging to one or more of the following groups:
 Lead, Credible Set, Consensus.}
    \item{\code{labels_subset}}{Include colored shapes and RSID labels
to help highlight SNPs belonging to one or more of the following groups:
 Lead, Credible Set, Consensus.}
    \item{\code{xtext}}{Include x-axis title and text for each track
(not just the lower-most one).}
    \item{\code{show_legend_genes}}{Show the legend for the \code{gene_track}.}
    \item{\code{zoom_exceptions_str}}{Names of tracks to exclude when zooming.}
    \item{\code{xgr_libnames}}{Passed to \link[echoplot]{XGR_plot}.
Which XGR annotations to check overlap with.
For full list of libraries see
\href{http://XGR_r-forge.r-project.org/#annotations-at-the-genomic-region-level}{
 here}.
 Passed to the \code{RData.customised} argument in \link[XGR]{xRDataLoader}.
 Examples:
 \itemize{
 \item{"ENCODE_TFBS_ClusteredV3_CellTypes"}
 \item{"ENCODE_DNaseI_ClusteredV3_CellTypes"}
 \item{"Broad_Histone"}
 }}
    \item{\code{xgr_n_top}}{Passed to \link[echoplot]{XGR_plot}.
Number of top annotations to be plotted
(passed to \link[echoannot]{XGR_filter_sources} and then 
\link[echoannot]{XGR_filter_assays}).}
    \item{\code{nott_epigenome}}{Include tracks showing brain cell-type-specific 
epigenomic data from 
\href{https://doi.org/10.1126/science.aay0793}{Nott et al. (2019)}.}
    \item{\code{nott_regulatory_rects}}{Include track generated by 
\link[echoannot]{NOTT2019_epigenomic_histograms}.}
    \item{\code{nott_show_placseq}}{Include track generated by
\link[echoannot]{NOTT2019_plac_seq_plot}.}
    \item{\code{nott_binwidth}}{When including Nott et al. (2019)
epigenomic data in the track plots,
adjust the bin width of the histograms.}
    \item{\code{nott_bigwig_dir}}{Instead of pulling Nott et al. (2019)
 epigenomic data
from the \emph{UCSC Genome Browser}, use a set of local bigwig files.}
    \item{\code{roadmap}}{Find and plot annotations from Roadmap.}
    \item{\code{roadmap_n_top}}{Passed to \link[echoplot]{ROADMAP_plot}.
Number of top annotations to be plotted
(passed to \link[echoannot]{ROADMAP_query}).}
    \item{\code{roadmap_query}}{Only plot annotations from Roadmap whose
metadata contains a string or any items from  a list of strings
(e.g. \code{"brain"} or \code{c("brain","liver","monocytes")}).}
    \item{\code{credset_thresh}}{The minimum fine-mapped posterior probability
 for a SNP to be considered part of a Credible Set.
For example, \code{credset_thresh=.95} means that all Credible Set SNPs
will be 95\% Credible Set SNPs.}
    \item{\code{facet_formula}}{Formula to facet plots by.
See \link[ggplot2]{facet_grid} for details.}
    \item{\code{save_plot}}{Save plot as RDS file.}
    \item{\code{plot_format}}{Format to save plot as 
when saving with \link[ggplot2]{ggsave}.}
    \item{\code{save_RDS}}{Save the tracks as an RDS file 
(\emph{Warning}: These plots take up a lot disk space).}
    \item{\code{return_list}}{Return a named list with each track as a separate plot
(default: \code{FALSE}). If \code{TRUE}, will return a merged plot using
\link[patchwork]{wrap_plots}.}
    \item{\code{dat}}{Data to query transcripts with.}
    \item{\code{LD_reference}}{LD reference to use:
\itemize{
\item{"1KGphase1" : }{1000 Genomes Project Phase 1 (genome build: hg19).}
\item{"1KGphase3" : }{1000 Genomes Project Phase 3 (genome build: hg19).}
\item{"UKB" : }{Pre-computed LD from a British
European-decent subset of UK Biobank.
\emph{Genome build} : hg19}
\item{"<vcf_path>" : }{User-supplied path to a custom VCF file 
to compute LD matrix from.\cr
\emph{Accepted formats}: \emph{.vcf} / \emph{.vcf.gz} / \emph{.vcf.bgz}\cr
\emph{Genome build} : defined by user with \code{target_genome}.}
\item{"<matrix_path>" : }{User-supplied path to a pre-computed LD matrix   
\emph{Accepted formats}:  \emph{.rds} / \emph{.rda} / \emph{.csv} / 
\emph{.tsv} / \emph{.txt}\cr
\emph{Genome build} : defined by user with \code{target_genome}.}
}}
    \item{\code{conda_env}}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}
    \item{\code{nThread}}{Number of threads to parallelize over.}
    \item{\code{LD_matrix}}{LD matrix.}
    \item{\code{zoom}}{Zoom into the center of the locus when plotting
(without editing the fine-mapping results file).
You can provide either:
\itemize{
\item{The size of your plot window in terms of basepairs
(e.g. \code{zoom=50000} for a 50kb window)}.
\item{How much you want to zoom in (e.g. \code{zoom="1x"}
for the full locus, \code{zoom="2x"}
for 2x zoom into the center of the locus, etc.)}.
}
You can pass a list of window sizes (e.g. \code{c(50000,100000,500000)})
to automatically generate
multiple views of each locus.
This can even be a mix of different style inputs: e.g.
 \code{c("1x","4.5x",25000)}.}
    \item{\code{genomic_units}}{Which genomic units to return window limits in.}
    \item{\code{density_adjust}}{Passed to \code{adjust} argument in 
\link[ggplot2]{geom_density}.}
    \item{\code{strip.text.y.angle}}{Angle of the y-axis facet labels.}
    \item{\code{dpi}}{dpi to use for raster graphics}
    \item{\code{height}}{height (defaults to the height of current
  plotting window)}
    \item{\code{width}}{width (defaults to the width of current
  plotting window)}
    \item{\code{consensus_thresh}}{The minimum number of fine-mapping tools 
in which a SNP is in the Credible Set in order to be included in the 
"Consensus_SNP" column.}
  }}
}
\description{
Plot multi-GWAS/QTL or multi-ancestry(i.e. trans-ethnic) fine-mapping
 results generated by tools like \link[echofinemap]{PAINTOR}.
}
\examples{
locus_dir <- file.path(tempdir(),echodata::locus_dir)
#### Make dat_ls ####
dat <- echodata::filter_snps(echodata::BST1, bp_distance = 10000) 
dat_ls <- list(gwas1=dat, gwas2=dat)
#### Make LD_ls ####
LD_matrix <- echodata::BST1_LD_matrix
LD_ls <- list(ancestry1=LD_matrix, ancestry2=LD_matrix)
#### Make plot ####
plot_list <- plot_locus_multi(dat_ls = dat_ls, 
                              LD_ls = LD_ls,
                              locus_dir = locus_dir)
}
