% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROADMAP_plot.R
\name{ROADMAP_plot}
\alias{ROADMAP_plot}
\title{Plot annotation peaks: ROADMAP}
\usage{
ROADMAP_plot(
  dat,
  roadmap_query,
  lib_name = "Roadmap.ChromatinMarks_CellTypes",
  locus_dir = tempdir(),
  limit_files = NULL,
  n_top = 5,
  adjust = 0.2,
  force_new = FALSE,
  show_plot = FALSE,
  conda_env = "echoR_mini",
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{dat}{\link[data.table]{data.table} with at least the following columns:
\describe{
\item{SNP}{SNP RSID}
\item{CHR}{chromosome}
\item{POS}{position}
}}

\item{roadmap_query}{Search all columns in the Roadmap annotations metadata
and only query annotations that contain your keywords.
Can provide multiple keywords in list form:
\code{c("placenta","liver","monocytes")}}

\item{lib_name}{Name of the data library to use.}

\item{locus_dir}{Locus-specific directory.}

\item{limit_files}{Limit the number of annotation files queried
(for faster testing).}

\item{n_top}{Number of top annotations to be plotted
(passed to \link[echoannot]{ROADMAP_query}).}

\item{adjust}{The granularity of the peaks.}

\item{force_new}{Download and prepare a new query
even if the file already exists locally (Default: \code{FALSE}).}

\item{show_plot}{Print the plot.}

\item{conda_env}{Conda environment to search for tabix in.}

\item{nThread}{Number of threads 
to parallelise downloading annotations over.}

\item{verbose}{Print messages.}
}
\value{
A named list containing:
\itemize{
\item{"data"}{\code{GRanges} object within the queried coordinates.}
\item{"plot"}{\code{ggbio} plot.}
}
}
\description{
Import annotation from \emph{ROADMAP}, filter to only those within the range
of \code{dat}, and then plot the peaks.
If the annotation has already been downloaded previously, it will be reused.
}
\examples{
dat <- echodata::BST1[seq_len(1000),]
roadmap_out <- echoplot::ROADMAP_plot(
    dat = dat,
    roadmap_query = "monocyte")
}
