% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROADMAP_track_plot.R
\name{ROADMAP_track_plot}
\alias{ROADMAP_track_plot}
\title{Plot ROADMAP query}
\source{
\code{
gr.snp <- echodata::dt_to_granges(echodata::BST1)
grl.roadmap <- echoannot::ROADMAP_query(
    query_dat = gr.snp,
    merge_and_process = TRUE,
    keyword_query = "monocyte")
track.roadmap <- ROADMAP_track_plot(grl.roadmap,
    gr.snp = gr.snp)
}
}
\usage{
ROADMAP_track_plot(
  grl.roadmap,
  gr.snp = NULL,
  geom = "density",
  adjust = 0.2,
  position = c("stack", "fill", "dodge"),
  show_plot = TRUE,
  as_ggplot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{grl.roadmap}{ROADMAP query results.}

\item{gr.snp}{Optionally, can include an extra \link[GenomicRanges]{GRanges}
object to ensure the plot does not extend beyond certain coordinates.}

\item{geom}{The type of plot to create.
Options include "density" and "histogram".}

\item{adjust}{The granularity of the peaks.}

\item{show_plot}{Whether to print the plot.}
}
\description{
Plot ROADMAP query
}
\keyword{internal}
