% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_track_merged_label.R
\name{snp_track_merged_label}
\alias{snp_track_merged_label}
\title{Add SNP labels}
\usage{
snp_track_merged_label(
  snp_plot,
  yvar = "PP",
  genomic_units = "Mb",
  show.legend = TRUE,
  max.overlaps = 30,
  verbose = TRUE
)
}
\arguments{
\item{genomic_units}{Which genomic units to return window limits in.}

\item{verbose}{Print messages.}
}
\description{
Add SNP labels for Manhattan plot.
Support function for \link[echoplot]{plot_locus}.
}
\keyword{internal}
