% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_model_track.R
\name{transcript_model_track}
\alias{transcript_model_track}
\title{Plot gene/transcript models}
\source{
\href{https://bioconductor.org/packages/release/bioc/vignettes/ensembldb/inst/doc/ensembldb.html}{ensembld tutorial}
\href{https://bioconductor.org/packages/devel/bioc/vignettes/Gviz/inst/doc/Gviz.html#45_GeneRegionTrack}{Gvix tutorial}
\href{http://bioconductor.org/packages/devel/bioc/vignettes/ggbio/inst/doc/ggbio.pdf}{ggbio tutorial}
}
\usage{
transcript_model_track(
  dat,
  max_transcripts = 1,
  remove_pseudogenes = TRUE,
  show.legend = TRUE,
  show_plot = FALSE,
  fill = "skyblue",
  shape = c("arrow", "box", "ellipse", "smallArrow"),
  transcriptAnnotation = c("symbol", "transcript"),
  collapseTranscripts = c(FALSE, TRUE, "longest"),
  stacking = c("squish", "hide", "dense", "pack", "full"),
  method = "ggplot",
  xtext = TRUE,
  expand_x_mult = c(0.05, 0.1),
  verbose = TRUE
)
}
\description{
Plot gene/transcript models
}
\examples{
data("LRRK2")
gene_track <- transcript_model_track(dat=LRRK2)
}
