% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR_plot.R
\name{XGR_plot}
\alias{XGR_plot}
\title{Plot annotation peaks: XGR}
\usage{
XGR_plot(
  dat,
  lib_name = "ENCODE_TFBS_ClusteredV3_CellTypes",
  locus_dir = tempdir(),
  palette = get_palettes(n_pals = 1, names_only = TRUE),
  fill_var = "Assay",
  facet_var = "Source",
  geom = "density",
  n_top = 5,
  adjust = 0.2,
  force_new = FALSE,
  show_plot = FALSE,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{dat}{\link[data.table]{data.table} with at least the following columns:
\describe{
\item{SNP}{SNP RSID}
\item{CHR}{chromosome}
\item{POS}{position}
}}

\item{lib_name}{Which XGR annotations to check overlap with.
For full list of libraries see
\href{http://XGR_r-forge.r-project.org/#annotations-at-the-genomic-region-level}{
 here.}
 Passed to the \code{RData.customised} argument in \link[XGR]{xRDataLoader}.}

\item{locus_dir}{Locus-specific directory.}

\item{palette}{Any palette available in \pkg{pals}.
See \link[echoplot]{get_palette} for a list of all palettes.}

\item{fill_var}{Fill variable.}

\item{facet_var}{Row facet variable.}

\item{geom}{Plot type ("density", or "histogram").}

\item{n_top}{Number of top annotations to be plotted
(passed to \link[echoannot]{XGR_filter_sources} and then
\link[echoannot]{XGR_filter_assays}).}

\item{adjust}{The granularity of the peaks.}

\item{force_new}{Download and prepare a new query
even if the file already exists locally (Default: \code{FALSE}).}

\item{show_plot}{Print the plot.}

\item{nThread}{Number of threads 
to parallelise downloading annotations over.}

\item{verbose}{Print messages.}
}
\value{
List with the "data" and the "plot".
}
\description{
Import annotation from \emph{XGR}, filter to only those within the range of
\code{dat}, and then plot the peaks.
If the annotation has already been downloaded previously, it will be reused.
}
\examples{
xgr_out <- echoplot::XGR_plot(dat = echodata::BST1[seq_len(1000),])
}
