% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_snp_labels.R
\name{construct_snp_labels}
\alias{construct_snp_labels}
\title{Construct SNP labels
 
Construct SNP labels for Manhattan plot.
Support function for \link[echoplot]{plot_locus}.}
\usage{
construct_snp_labels(
  dat,
  labels_subset = c("Lead", "CS", "UCS", "Consensus"),
  remove_duplicates = TRUE,
  grouping_vars = c("SNP"),
  merge_with_input = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dat}{Data to query transcripts with.}

\item{verbose}{Print messages.}
}
\description{
Construct SNP labels
 
Construct SNP labels for Manhattan plot.
Support function for \link[echoplot]{plot_locus}.
}
\keyword{internal}
