% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transcripts.R
\name{get_transcripts}
\alias{get_transcripts}
\title{Get transcripts}
\source{
\code{
#### Alternative approaches I've tried ####
dat <- echodata::LRRK2
gr.snp <- echodata::dt_to_granges(dat = dat)

# Warning:: MUST load the full package bc
# it loads other necessary packages into the namespace.

BiocManager::install("Homo.sapiens")
library(Homo.sapiens)
txdb <- Homo.sapiens::Homo.sapiens
GenomicFeatures::genes(txdb)

BiocManager::install("TxDb.Hsapiens.UCSC.hg19.knownGene")
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene
GenomicFeatures::genes(txdb)
}
}
\usage{
get_transcripts(
  gr.snp,
  max_transcripts = 1,
  remove_pseudogenes = TRUE,
  tx_biotypes = NULL,
  verbose = TRUE
)
}
\arguments{
\item{max_transcripts}{Maximum number of transcripts per gene.}

\item{tx_biotypes}{Transcript biotypes to include in the gene model track. 
By default (\code{NULL}), all transcript biotypes will be included.
See \link[echoplot]{get_transcripts_biotypes} for a full list of 
all available biotypes}

\item{verbose}{Print messages.}
}
\description{
Get transcript positions for all genes in a range.
}
\keyword{internal}
