% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_window_limits.R
\name{get_window_limits}
\alias{get_window_limits}
\title{Get window limits}
\usage{
get_window_limits(
  dat,
  index_as_center = TRUE,
  plot_zoom = NULL,
  genomic_units = "Mb",
  verbose = TRUE
)
}
\description{
Get window size limits for Manhattan plot.
}
\examples{
dat <- echodata::BST1
xlims <- get_window_limits(dat=dat, plot_zoom=50000)
xlims <- get_window_limits(dat=dat, plot_zoom="all")
xlims <- get_window_limits(dat=dat, plot_zoom="5x")
}
\seealso{
Other plot: 
\code{\link{add_multitrack_lines}()},
\code{\link{dot_summary_plot}()},
\code{\link{get_window_suffix}()}
}
\concept{plot}
\keyword{internal}
