% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_locus.R
\name{plot_locus}
\alias{plot_locus}
\title{Generate a locus plot}
\usage{
plot_locus(
  dat,
  locus_dir,
  LD_matrix = NULL,
  LD_reference = NULL,
  facet_formula = "Method~.",
  dataset_type = "GWAS",
  color_r2 = TRUE,
  finemap_methods = c("ABF", "FINEMAP", "SUSIE", "POLYFUN_SUSIE"),
  track_order = NULL,
  track_heights = NULL,
  plot_full_window = TRUE,
  dot_summary = FALSE,
  qtl_suffixes = NULL,
  mean.PP = TRUE,
  credset_thresh = 0.95,
  consensus_thresh = 2,
  sig_cutoff = 5e-08,
  gene_track = TRUE,
  tx_biotypes = NULL,
  point_size = 1,
  point_alpha = 0.6,
  density_adjust = 0.2,
  snp_group_lines = c("Lead", "UCS", "Consensus"),
  xtext = FALSE,
  show_legend_genes = TRUE,
  xgr_libnames = NULL,
  xgr_n_top = 5,
  roadmap = FALSE,
  roadmap_query = NULL,
  roadmap_n_top = 7,
  zoom_exceptions_str = "*full window$|zoom_polygon",
  nott_epigenome = FALSE,
  nott_regulatory_rects = TRUE,
  nott_show_placseq = TRUE,
  nott_binwidth = 200,
  nott_bigwig_dir = NULL,
  save_plot = FALSE,
  show_plot = TRUE,
  genomic_units = "Mb",
  strip.text.y.angle = 0,
  max_transcripts = 1,
  zoom = c("1x"),
  dpi = 300,
  height = 12,
  width = 10,
  plot_format = "png",
  save_RDS = FALSE,
  return_list = FALSE,
  conda_env = "echoR_mini",
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{dat}{Data to query transcripts with.}

\item{locus_dir}{Storage directory to use.}

\item{LD_matrix}{LD matrix.}

\item{LD_reference}{LD reference to use:
\itemize{
\item{"1KGphase1" : }{1000 Genomes Project Phase 1 (genome build: hg19).}
\item{"1KGphase3" : }{1000 Genomes Project Phase 3 (genome build: hg19).}
\item{"UKB" : }{Pre-computed LD from a British
European-decent subset of UK Biobank (genome build: hg19).}
\item{"<vcf_path>" : }{User-supplied path to a custom VCF file 
to compute LD matrix from 
(genome build: defined by user with \code{target_genome}).}
}}

\item{dataset_type}{Dataset type (e.g. "GWAS" or "eQTL").}

\item{color_r2}{Whether to color data points (SNPs)
by how strongly they correlate with the lead SNP 
(i.g. LD measured in terms of r2).}

\item{finemap_methods}{Fine-mapping methods to plot tracks for, 
where the y-axis show the Posterior Probabilities (PP) 
of each SNP being causal.}

\item{track_order}{The order in which tracks should appear
(from top to bottom).}

\item{track_heights}{The height of each track (from top to bottom).}

\item{plot_full_window}{Include a track with a Manhattan plot of the full
GWAS/eQTL locus (not just the zoomed-in portion).}

\item{dot_summary}{Include a dot-summary plot that highlights the
Lead, Credible Set, and Consensus SNPs.}

\item{qtl_suffixes}{If columns with QTL data is included in \code{dat},
you can indicate which columns those are with one or more string suffixes
(e.g. \code{qtl_suffixes=c(".eQTL1",".eQTL2")} to use the columns 
"P.QTL1", "Effect.QTL1", "P.QTL2", "Effect.QTL2").}

\item{mean.PP}{Include a track showing mean Posterior Probabilities (PP) 
averaged across all fine-mapping methods.}

\item{credset_thresh}{The minimum fine-mapped posterior probability
 for a SNP to be considered part of a Credible Set.
For example, \code{credset_thresh=.95} means that all Credible Set SNPs
will be 95\% Credible Set SNPs.}

\item{consensus_thresh}{The minimum number of fine-mapping tools 
in which a SNP is in the Credible Set in order to be included in the 
"Consensus_SNP" column.}

\item{sig_cutoff}{Filters out SNPs to plot based on an (uncorrected)
p-value significance cutoff.}

\item{gene_track}{Include a track showing gene bodies.}

\item{tx_biotypes}{Transcript biotypes to include in the gene model track. 
By default (\code{NULL}), all transcript biotypes will be included.
See \link[echoplot]{get_transcripts_biotypes} for a full list of 
all available biotypes}

\item{point_size}{Size of each data point.}

\item{point_alpha}{Opacity of each data point.}

\item{density_adjust}{Passed to \code{adjust} argument in 
\link[ggplot2]{geom_density}.}

\item{snp_group_lines}{Include vertical lines to help highlight 
SNPs belonging to one or more of the following groups:
 Lead, Credible Set, Consensus.}

\item{xtext}{Include x-axis title and text for each track
(not just the lower-most one).}

\item{show_legend_genes}{Show the legend for the \code{gene_track}.}

\item{xgr_libnames}{Passed to \link[echoannot]{XGR_plot}.
Which XGR annotations to check overlap with.
For full list of libraries see
\href{http://XGR_r-forge.r-project.org/#annotations-at-the-genomic-region-level}{
 here}.
 Passed to the \code{RData.customised} argument in \link[XGR]{xRDataLoader}.}

\item{xgr_n_top}{Passed to \link[echoannot]{XGR_plot}.
Number of top annotations to be plotted
(passed to \link[echoannot]{XGR_filter_sources} and then 
\link[echoannot]{XGR_filter_assays}).}

\item{roadmap}{Find and plot annotations from Roadmap.}

\item{roadmap_query}{Only plot annotations from Roadmap whose
metadata contains a string or any items from  a list of strings
(e.g. \code{"brain"} or \code{c("brain","liver","monocytes")}).}

\item{roadmap_n_top}{Passed to \link[echoannot]{ROADMAP_plot}.
Number of top annotations to be plotted
(passed to \link[echoannot]{ROADMAP_merge_and_process_grl}).}

\item{zoom_exceptions_str}{Names of tracks to exclude when zooming.}

\item{nott_epigenome}{Include tracks showing brain cell-type-specific 
epigenomic data from 
\href{https://doi.org/10.1126/science.aay0793}{Nott et al. (2019)}.}

\item{nott_regulatory_rects}{Include track generated by 
\link[echoannot]{NOTT2019_epigenomic_histograms}.}

\item{nott_show_placseq}{Include track generated by
\link[echoannot]{NOTT2019_plac_seq_plot}.}

\item{nott_binwidth}{When including Nott et al. (2019)
epigenomic data in the track plots,
adjust the bin width of the histograms.}

\item{nott_bigwig_dir}{Instead of pulling Nott et al. (2019)
 epigenomic data
from the \emph{UCSC Genome Browser}, use a set of local bigwig files.}

\item{save_plot}{Save plot as RDS file.}

\item{show_plot}{Print plot to screen.}

\item{genomic_units}{Which genomic units to return window limits in.}

\item{strip.text.y.angle}{Angle of the y-axis facet labels.}

\item{max_transcripts}{Maximum number of transcripts per gene.}

\item{zoom}{Zoom into the center of the locus when plotting
(without editing the fine-mapping results file).
You can provide either:
\itemize{
\item{The size of your plot window in terms of basepairs
(e.g. \code{zoom=50000} for a 50kb window)}.
\item{How much you want to zoom in (e.g. \code{zoom="1x"}
for the full locus, \code{zoom="2x"}
for 2x zoom into the center of the locus, etc.)}.
}
You can pass a list of window sizes (e.g. \code{c(50000,100000,500000)})
to automatically generate
multiple views of each locus.
This can even be a mix of different style inputs: e.g.
 \code{c("1x","4.5x",25000)}.}

\item{dpi}{dpi to use for raster graphics}

\item{height}{height (defaults to the height of current
  plotting window)}

\item{width}{width (defaults to the width of current
  plotting window)}

\item{plot_format}{Format to save plot as 
when saving with \link[ggplot2]{ggsave}.}

\item{save_RDS}{Save the tracks as an RDS file 
(\emph{Warning}: These plots take up a lot disk space).}

\item{return_list}{Return a named list with each track as a separate plot
(default: \code{FALSE}). If \code{TRUE}, will return a merged plot using
\link[patchwork]{wrap_plots}.}

\item{conda_env}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}

\item{nThread}{Number of threads to parallelize over.}

\item{verbose}{Print messages.}
}
\description{
Generate a locus-specific plot with multiple selectable tracks.
Users can also generate multiple zoomed in views of the plot at
 multiple resolutions.
}
\examples{
dat<- echodata::BST1
LD_matrix <- echodata::BST1_LD_matrix
locus_dir <- file.path(tempdir(),echodata::locus_dir)

plt <- echoplot::plot_locus(dat = dat,
                            locus_dir = locus_dir,
                            LD_matrix = LD_matrix,
                            show_plot = FALSE)
}
