% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_track_merged.R
\name{snp_track_merged}
\alias{snp_track_merged}
\title{SNP track merged}
\usage{
snp_track_merged(
  dat,
  yvar = "-log10(P)",
  labels_subset = c("Lead", "UCS", "Consensus"),
  absolute_labels = FALSE,
  label_type = "rsid_only",
  sig_cutoff = 5e-08,
  cutoff_lab = paste("P <", sig_cutoff),
  point_alpha = 0.5,
  show.legend = TRUE,
  xtext = TRUE,
  facet_formula = "Method~.",
  dataset_type = NULL,
  genomic_units = "POS",
  strip.text.y.angle = 0,
  show_plot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dat}{GWAS summary statistics subset to query the LD panel with.}

\item{verbose}{Print messages.}
}
\description{
Manhattan plot of GWAS/QTL data with various fine-mapping 
related annotations. Support function for \link[echoplot]{plot_locus}.
}
\keyword{internal}
