% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_track_merged.R
\name{snp_track_merged}
\alias{snp_track_merged}
\title{SNP track merged}
\usage{
snp_track_merged(
  dat,
  yvar = "-log10(P)",
  labels_subset = c("Lead", "UCS", "Consensus"),
  absolute_labels = FALSE,
  label_type = "rsid_only",
  label_leadsnp = TRUE,
  sig_cutoff = 5e-08,
  point_alpha = 0.5,
  show.legend = TRUE,
  xtext = TRUE,
  facet_formula = "Method~.",
  dataset_type = NULL,
  genomic_units = "POS",
  strip.text.y.angle = 0,
  show_plot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dat}{Data to query transcripts with.}

\item{sig_cutoff}{Filters out SNPs to plot based on an (uncorrected)
p-value significance cutoff.}

\item{point_alpha}{Opacity of each data point.}

\item{xtext}{Include x-axis title and text for each track
(not just the lower-most one).}

\item{dataset_type}{Dataset type (e.g. "GWAS" or "eQTL").}

\item{genomic_units}{Which genomic units to return window limits in.}

\item{strip.text.y.angle}{Angle of the y-axis facet labels.}

\item{show_plot}{Print plot to screen.}

\item{verbose}{Print messages.}
}
\description{
Manhattan plot of GWAS/QTL data with various fine-mapping 
related annotations. Support function for \link[echoplot]{plot_locus}.
}
\examples{
dat <- echodata::BST1[seq_len(100),]
plt <- snp_track_merged(dat = dat)
}
\keyword{internal}
