% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_model_track.R
\name{transcript_model_track}
\alias{transcript_model_track}
\title{Plot transcripts}
\source{
\href{https://bioconductor.org/packages/release/bioc/vignettes/ensembldb/inst/doc/ensembldb.html}{ensembld tutorial}
\href{https://bioconductor.org/packages/devel/bioc/vignettes/Gviz/inst/doc/Gviz.html#45_GeneRegionTrack}{Gvix tutorial}
\href{http://bioconductor.org/packages/devel/bioc/vignettes/ggbio/inst/doc/ggbio.pdf}{ggbio tutorial}
}
\usage{
transcript_model_track(
  dat,
  max_transcripts = 1,
  tx_biotypes = NULL,
  remove_pseudogenes = TRUE,
  show_plot = FALSE,
  show.legend = FALSE,
  fill = "skyblue",
  shape = c("arrow", "box", "ellipse", "smallArrow"),
  transcriptAnnotation = c("symbol", "transcript"),
  collapseTranscripts = c(FALSE, TRUE, "longest"),
  stacking = c("squish", "hide", "dense", "pack", "full"),
  method = "ggplot",
  xtext = TRUE,
  expand_x_mult = c(0.05, 0.1),
  verbose = TRUE
)
}
\arguments{
\item{dat}{Data to query transcripts with.}

\item{max_transcripts}{Maximum number of transcripts per gene.}

\item{tx_biotypes}{Transcript biotypes to include in the gene model track. 
By default (\code{NULL}), all transcript biotypes will be included.
See \link[echoplot]{get_tx_biotypes} for a full list of 
all available biotypes}

\item{remove_pseudogenes}{Whether to remove known pseudogenes.}

\item{show_plot}{Print plot to screen.}

\item{show.legend}{Show gene plot legend.}

\item{fill}{A variable name string, or a color to fill the plot data with.}

\item{shape}{Shape type to plot genes as. 
Passed to \link[Gviz]{plotTracks}.}

\item{transcriptAnnotation}{Whether to show the gene symbol
 or the full transcript ID.
Passed to \link[Gviz]{plotTracks}.}

\item{collapseTranscripts}{Logical or character scalar. 
Can be one in gene, longest, shortest or meta. 
Merge all transcripts of the same gene into one single gene model. 
In the case of gene (or TRUE), this will only keep the start location 
of the first exon and the end location of the last exon from all 
transcripts of the gene. For shortest and longest, only the longest 
or shortest transcript model is retained. For meta, a meta-transcript
 containing the union of all exons is formed 
 (essentially identical to the operation \code{reduce(geneModel)}).
Passed to \link[Gviz]{plotTracks}.}

\item{stacking}{The stacking type for overlapping items of the track. One in
\code{c(hide, dense, squish, pack,full)}. Currently, only hide (don't show
the track items, squish (make best use of the available space) and dense (no
stacking at all) are implemented.}

\item{method}{Method to use: "gviz" or "ggplot".}

\item{xtext}{Include x-axis title and text.}

\item{expand_x_mult}{Expand the x-axis limits to include partially 
overlapping transcripts.}

\item{verbose}{Print messages.}
}
\description{
Plot a max number of transcripts per genes.
See \link[echoplot]{get_tx_biotypes} for a full list of
all available biotypes.
}
\examples{
dat <- echodata::BST1 
gene_track <- echoplot::transcript_model_track(dat=dat)
}
